@@ -1,9 +1,9 @@
 {
   "name": "@angular-extensions/elements",
-  "version": "7.2.0",
+  "version": "8.9.0",
   "peerDependencies": {
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0"
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0"
   },
   "publishConfig": {
     "access": "public"
