@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { MechanismComponent } from './mechanism.component';
+import { ConfigurationComponent } from './configuration.component';
 
-const routes: Routes = [{ path: '', component: MechanismComponent }];
+const routes: Routes = [{ path: '', component: ConfigurationComponent }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class MechanismRoutingModule {}
+export class ConfigurationRoutingModule {}
