@@ -12,8 +12,14 @@ import {
   MatToolbarModule,
   MatGridListModule,
   MatExpansionModule,
-  MatSnackBarModule
+  MatSnackBarModule,
+  MatSliderModule,
+  MatFormFieldModule,
+  MatSlideToggleModule
 } from '@angular/material';
+import { SpinnerComponent } from './spinner/spinner.component';
+import { ErrorComponent } from './error/error.component';
+import { RootErrorComponent } from './root-error/root-error.component';
 
 const ANGULAR_MATERIAL_MODULES = [
   MatCardModule,
@@ -25,13 +31,16 @@ const ANGULAR_MATERIAL_MODULES = [
   MatSidenavModule,
   MatGridListModule,
   MatExpansionModule,
-  MatSnackBarModule
+  MatSnackBarModule,
+  MatSliderModule,
+  MatFormFieldModule,
+  MatSlideToggleModule
 ];
 
 @NgModule({
-  declarations: [],
+  declarations: [SpinnerComponent, ErrorComponent, RootErrorComponent],
   imports: [CommonModule, ...ANGULAR_MATERIAL_MODULES],
-  exports: [CommonModule, ...ANGULAR_MATERIAL_MODULES]
+  exports: [CommonModule, ...ANGULAR_MATERIAL_MODULES, SpinnerComponent]
 })
 export class SharedModule {
   constructor(
