@@ -4,7 +4,7 @@
 
 export const environment = {
   // version:start
-  version: '8.5.0',
+  version: '8.9.0',
   // version:end
   production: false
 };
