@@ -1,10 +1,17 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Type, Optional, Inject } from '@angular/core';
+
+import { LazyElementRootOptions } from './lazy-elements.module';
+import { LAZY_ELEMENT_ROOT_OPTIONS } from './lazy-elements.tokens';
 
 const LOG_PREFIX = '@angular-extensions/elements';
 
 export interface ElementConfig {
   tag: string;
   url: string;
+  isModule?: boolean;
+  loadingComponent?: Type<any>;
+  errorComponent?: Type<any>;
+  preload?: boolean;
 }
 
 @Injectable({
@@ -14,56 +21,86 @@ export class LazyElementsLoaderService {
   registry: Map<string, Promise<void>> = new Map<string, Promise<void>>();
   configs: ElementConfig[] = [];
 
-  constructor() {}
+  constructor(
+    @Optional()
+    @Inject(LAZY_ELEMENT_ROOT_OPTIONS)
+    public options: LazyElementRootOptions
+  ) {
+    if (!options) {
+      this.options = {};
+    }
+  }
 
   addConfigs(newConfigs: ElementConfig[]) {
     newConfigs.forEach(newConfig => {
-      const existingConfig = this.configs.find(
-        config => config.tag === newConfig.tag
-      );
+      const existingConfig = this.getElementConfig(newConfig.tag);
       if (existingConfig) {
         console.warn(
           `${LOG_PREFIX} - ElementConfig for tag '${newConfig.tag}' was previously added, it will not be added multiple times, continue...`
         );
       } else {
         this.configs.push(newConfig);
+        const shouldPreload =
+          newConfig.preload !== undefined
+            ? newConfig.preload
+            : this.options.preload;
+        if (shouldPreload) {
+          this.loadElement(newConfig.url, newConfig.tag, newConfig.isModule);
+        }
       }
     });
   }
 
-  loadElement(
-    url: string,
-    tag: string,
-    isModule: boolean = false
-  ): Promise<void> {
-    let elementUrl = url;
+  getElementConfig(tag: string): ElementConfig {
+    return this.configs.find(config => config.tag === tag);
+  }
+
+  preload(tags?: string[]) {
+    let configs = this.configs;
+    if (tags) {
+      configs = this.configs.filter(config => tags.includes(config.tag));
+    }
+    configs.forEach(config =>
+      this.loadElement(config.url, config.tag, config.isModule)
+    );
+  }
+
+  loadElement(url: string, tag: string, isModule?: boolean): Promise<void> {
+    const config = this.getElementConfig(tag);
+
     if (!url) {
-      const config = this.configs.find(c => c.tag === tag);
       if (!config) {
         throw new Error(`${LOG_PREFIX} - url for <${tag}> not found`);
       }
-      elementUrl = config.url;
+      url = config.url;
+    }
+
+    if (isModule === undefined) {
+      isModule =
+        config && config.isModule !== undefined
+          ? config.isModule
+          : this.options.isModule;
     }
 
     if (!tag) {
       throw new Error(
-        `${LOG_PREFIX} - tag for '${elementUrl}' not found, the *axLazyElement has to be used on HTML element`
+        `${LOG_PREFIX} - tag for '${url}' not found, the *axLazyElement has to be used on HTML element`
       );
     }
 
-    if (!this.hasElement(elementUrl)) {
-      const notifier = this.addElement(elementUrl);
+    if (!this.hasElement(url)) {
+      const notifier = this.addElement(url);
       const script = document.createElement('script') as HTMLScriptElement;
       if (isModule) {
         script.type = 'module';
       }
-      script.src = elementUrl;
+      script.src = url;
       script.onload = notifier.resolve;
       script.onerror = notifier.reject;
       document.body.appendChild(script);
     }
 
-    return this.registry.get(this.stripUrlProtocol(elementUrl));
+    return this.registry.get(this.stripUrlProtocol(url));
   }
 
   private addElement(url: string): Notifier {
